S81 : new structure for ProfilAidant section
